+++ CHANGELOG.md
@@ -5,10 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] - yyyy-mm-dd
+## [0.0.6-preview] - yyyy-mm-dd
 
 ### Added
-- 
+- New PR Template into the `/.github/` directory
 
 ### Changed
 - 
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 -
 
-## [0.0.5] - yyyy-mm-dd
+## [0.0.5] - 2025-02-24
 
 ### Added
 - Added missing `Circle2Circle()` function in [`CollisionChecks.cs`](Runtime/Scripts/Utilities/Physics/CollisionChecks.cs)
@@ -94,6 +94,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - 
 
 ### Fixed
+-
+
+## [0.0.1-6000] - 2024-12-16
+
+Updates to main branch to function on **Unity 6**
+
+### Changed
+- Updated target Unity version to `6000.0`
+- Removed packages that are no longer use with **Unity 6**
+  - TextmeshPro
+- Updated Cinemachine to `3.1.2`
+- Updated InputSystem to `1.11.2`
+- 
+
+### Fixed
+- Fixed namespace upgrade requirement for Cinemachine: `Cinemachine` -> `Unity.Cinemachine`
+- Fixed `CinemachineVirtualCamera` -> `CinemachineCamera`
+- Fixed `CinemachineSmoothPath` -> `SplineContainer`
 -
 
 ## [0.0.1] - 2024-12-15